Collect common translate and block_edit php code
No functional change intended.
