+++ userguide/translate.php
@@ -2,6 +2,8 @@
 require('inc/common.php');
 role_needed(ROLE_TRANSLATOR);
 
+require_once('inc/blocks.php');
+
 $xml_msg = '';
 
 if (isset($_POST['translate_doc']) and isset($_POST['translate_string'])
@@ -109,21 +111,8 @@ if (!$row)
 
 $lang_name = $row['lang_name'];
 
-$req = db_query('
-	SELECT path_original, path_translations FROM ' . DB_DOCS . "
-	WHERE doc_id = ?", array($doc_id));
-$row = db_fetch($req);
-db_free($req);
-
-if (!$row)
-	redirect('index.php');
-
-$path_original = $row['path_original'];
-$path_trans = dirname(str_replace('{LANG}', $lang, $row['path_translations']));
-
-$doc = new DOMDocument();
-$doc->load(REF_DIR . '/' . $path_original)
-	or die('Unable to load the document.');
+$doc = load_doc_with_blocks($doc_id, $lang);
+$head = $doc->getElementsByTagName('head')->item(0);
 
 // Get all the translations
 $col_name = 'translation_' . $lang;
@@ -134,52 +123,24 @@ $req = db_query('SELECT string_id, "' . $col_name  . '", "' . $col_fuzzy . '"
 	WHERE doc_id = ?" . '
 	ORDER BY string_id', array($doc_id));
 
-// Build the JavaScript array that will hold the source and translation strings
+// Build the JavaScript array that will hold the translation strings
 $js = '';
-$js .= 'const base_url = \'' . str_replace("'", "\'", $base_url) . "';\n";
-$js .= 'const base_local = \'' . str_replace("'", "\'", $path_trans) . "';\n";
 $js .= 'const lang = window.lang = \'' . str_replace("'", "\'", $lang) . "';\n";
 $js .= 'const lang_name = window.lang_name = \'' . str_replace("'", "\'", $lang_name) . "';\n";
-$js .= "const doc_id = $doc_id;\n";
-$js .= "var source_strings = new Array();\n";
 $js .= "var translated_strings = new Array();\n";
 $js .= "var is_fuzzy = new Array();\n";
 
 while ($row = db_fetch($req)) {
 	$id = $row['string_id'];
 
-	$translated = str_replace('\\', '\\\\', $row[$col_name]);
-	$translated = str_replace('"', '\"', $translated);
-	$translated = str_replace("\n", '\n', $translated);
-	$translated = str_replace("\r", '', $translated);
+	$translated = escape_js_string($row[$col_name]);
 
 	$js .= "translated_strings[$id] = \"$translated\";\n";
 	if ($row[$col_fuzzy])
 		$js .= "is_fuzzy[$id] = 1;\n";
 }
 
-$js .= get_source_strings($doc);
-
 db_free($req);
-
-$doc->loadXML(replace_placeholders($doc->saveXML(), $lang));
-
-$head = $doc->getElementsByTagName('head')->item(0);
-
-// Redirect all links of the page to the translated version
-$metas = $doc->getElementsByTagName('meta');
-$node = $head->firstChild;
-if ($metas) {
-	foreach ($metas as $meta)
-		if ($meta->getAttribute('http-equiv') == 'content-type')
-			$node = $meta;
-}
-
-$node = append_sibling($doc->createTextNode("\n\t"), $node);
-$base = $doc->createElement('base');
-$base->setAttribute('href', $base_url . '/' . EXPORT_DIR . '/' . $path_trans . '/');
-$node = append_sibling($base, $node);
-$node = append_sibling($doc->createTextNode("\t"), $node);
 
 // Pass the script path to JavaScript
 append_js_code($head, $js);
@@ -187,74 +148,7 @@ append_js_code($head, $js);
 // Include the JavaScript translation helper
 append_js_file($head, $base_url . '/shared/translate_tool.js');
 
-html_set_lang($doc, $lang);
-
 echo $doc->saveXML();
-
-function append_js_code($node, $code) {
-	global $doc;
-	// We are generating an XML document so the JS code should be encapsuled in a
-	// CDATA tag. However, the opening <![CDATA[ and closing ]]> must be preceded
-	// by "//" so they are treated like JS comments.
-
-	$js = $doc->createElement('script');
-	$js->setAttribute('type', 'text/JavaScript');
-
-	$js->appendChild($doc->createTextNode("\n//"));
-	$js->appendChild($doc->createCDATASection("\n$code\n//"));
-	$js->appendChild($doc->createTextNode("\n"));
-
-	$node->appendChild($js);
-	$node->appendChild($doc->createTextNode("\n"));
-}
-
-function append_js_file($node, $file) {
-	global $doc;
-
-	$js = $doc->createElement('script');
-	$js->setAttribute('type', 'text/JavaScript');
-	$js->setAttribute('src', $file);
-
-	$node->appendChild($js);
-	$node->appendChild($doc->createTextNode("\n"));
-}
-
-function get_source_strings($node) {
-	static $used_ids = array();
-	$to_return = '';
-	foreach ($node->childNodes as $child) {
-		if ($child instanceOf DOMElement) {
-			if ($child->hasAttribute(ATTR_TRANS_ID)) {
-				$id = intval($child->getAttribute(ATTR_TRANS_ID));
-				if (isset($used_ids[$id]))
-					continue;
-
-				$used_ids[$id] = true;
-
-				$text = DOMInnerHTML($child);
-
-				$text = str_replace('\\', '\\\\', $text);
-				$text = str_replace('"', '\"', $text);
-				$text = str_replace("\n", '\n', $text);
-
-				$to_return .= "source_strings[$id] = \"$text\";\n";
-			} else {
-				$to_return .= get_source_strings($child);
-			}
-		}
-	}
-
-	return $to_return;
-}
-
-function xml_error_handler($errno, $errstr, $errfile, $errline) {
-	global $xml_msg;
-
-	$err = (preg_match('/^DOMDocument::loadXML\(\) \[.*\]: (.*)$/',
-		$errstr, $matches) ? $matches[1] : $errstr);
-
-	$xml_msg .= htmlspecialchars_decode($err) . "\n";
-}
 
 function check_xml_tags($text_original, $text_translation) {
 	global $relaxed_parsing_attributes, $relaxed_parsing_complete;
